ray function was taking float instead of double, freeing xs.intersections on fail, multithread function wasnt setting error to true on pthread failure, and i was initializing thread count twice pointlessly

+++ includes/rays.h
@@ -6,7 +6,7 @@
 /*   By: bpetrovi <bpetrovi@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/30 18:56:37 by bpetrovi          #+#    #+#             */
-/*   Updated: 2026/06/09 20:59:33 by bpetrovi         ###   ########.fr       */
+/*   Updated: 2026/08/23 18:48:32 by bpetrovi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ typedef struct s_ray
 
 t_ray		r_init(t_tuple origin, t_tuple direction);
 t_ray		r_transform(t_ray ray, t_matrix transformation);
-t_tuple		r_pos(t_ray ray, float time);
+t_tuple		r_pos(t_ray ray, double time);
 
 #endif
